pyproject.toml
[project]
name = "gofreight-homework"
version = "0.1.0"
description = "Natural-language Data Commons query CLI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pydantic>=2.11.0",
  "python-dotenv>=1.1.0",
  "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
dc-query = "dc_nl_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"

[tool.hatch.build.targets.wheel]
packages = ["src/dc_nl_cli"]

[dependency-groups]
dev = [
    "pre-commit>=4.6.0",
    "ruff>=0.15.12",
]
